added Echo! on by default; will add menu option soon.

+++ sound.c
@@ -32,17 +32,18 @@ UINT getFrequencies(int note, int bend){
 /* Play frequency to channel 1 and 2.
  * Frequencies are in non-hertz values.
  */
-void play_freq_ch1 (UINT freq, int newNote )
+void play_freq_ch1 (UINT freq, UINT8 newNote )
 {
     NR13_REG = (unsigned char) freq;
     if (newNote == 1) {
         NR14_REG = 0x80 | (freq >> 8);
+         
     } else { 
         NR14_REG = 0x00 | (freq >> 8);
     }
 }
 
-void play_freq_ch2 (UINT freq, int newNote )
+void play_freq_ch2 (UINT freq, UINT8 newNote )
 {
     NR23_REG = (unsigned char) freq;
